Move function to radio field
